refactor(tests): move configuration and rule-pack tests into sibling modules
